Warn the user of UPnP missing dependencies only if he wants to use it.
The user will be warn of missing UPnP dependencies only if the config
wants UPnP to be active.
Since the user can't activate UPnP via the GUI if the dependencies are missing,
the only case where he will be warned is:
The user has accidentally removes the UPnP dependencies (system reinstall or
package management screw up) and UPnP is still active in the config.