Allow directories to be selected by a shortcut from the gtkfilechooserbutton
The wrong event was used in the glade file current_folder_changed instead of selection_changed.
This was preventing the change of downloaddir, uploaddir and incompletedir by direct
selection of a gtk shortcut without opening the FileChooserDialog.

Also use Gio to avoid using the recode function.
Ideally we should use Gio everywhere where must work with files or directories.