Fix for selecting position of tabs in the settings
When the UI was not in english the positions (top, bottom, left, right) of tabs
you can choose in the settings window were translated and store translated in the config file.

The tab positionning need a GtkPositionType which is either "top", "bottom", "left" or "right".
Since these labels were translated positionning didn't work.

The solution is :
* Always store a GtkPositionType in the config.
* Use a GtkComboBoxEntry in the settings and bind it with a Gtk ListStore of two columns.
  The first column is translated and displayed.
  The second is hidden and contain the corresponding GtkPositionType.