Simply server config in the settings window
The server that should be used is server.slsknet.org:2242.
The server/port server.slsknet.org:2240 was a test network that's deprecated for a while now.

The user can still customise the server/port if he really wants to
via the server -> server configuration option in the n+ config file.