Drop OSX packaging
There's no one to maintain the macOS port of N+.
It might be possible to make it work with brew or macports but the helper scripts have been unmaintained for ages.
If someone would like to step up and maintain it, we would be happy to help him include new build scripts.