Fix connections problems (upload & download)
An internal array "peerconns" maintaining users connections was keeping objects
(users connections info) without any valid connection object in it.
It was trying to reuse an not existant connection to send messages to peers.
These connections were closed either via a Timer (searchResultsTimer in pynicotine.py)
or via another call to ClosePeerConnection in the function FileSearchResult.

For the time being we recreate the connection but these two functions are duplicate
in a sense that they do the same thing and might cause a race condition.
One of them should be removed to avoid this.

Also the "peerconns" array should be replaced in the future by a dict to increase lookup speed
and avoid doing lots of loops everywhere to identify a user connection.