shares: some files were not included in the filestream
Even with a rebuild some files where not included in the filestream.
The filestream is the packed datastructure sent to peer that include our file list.

For exemple let's say we have a file in a directory previously shared but it's metadata aren't
handle correctly.
If we fix a bug or update mutagen to fix metadata handling, the file still has the same timestamp
and it's directory too.
The file info was actually rebuild but not the filestream associated: fix that.