Drop the ReqidManager: was failing to track transfers
The ReqidManager was failing to track transfers and leaking memory.
Tracking transfers need to be rewritten one of these days to avoid loops.
For the time being the overhead is minimal from a CPU perpective.
It's the first part of getting rid of a lot of hacks in temporary.py.