Shares: only check if last folder in path is hidden, not every folder
The previous method of checking every folder in a path is flawed, since it results in us being unable to share any folders having >=1 parent folders starting with a dot. This can be observed when attempting to run the shares unit test in autopkgtest, since the test files are located inside a ".pybuild" folder.

We should only check if subfolders of a shared folder are hidden.