search.ui: Layout tweaks and hints
Changed: Layout of Search History options, to group the widgets together.
Added: Show privately shared files option, because it is misplaced in userinterface.ui
Changed: Layout of Search Result Filter options, because the Enable default filters toggle was in a misleading place. 
Changed: Position and label of "Free slot" -> Longer string similar to tooltip and Result Filter Help.
Changed: Position of Include and Exclude, because these have minimal purpose to set as defaults here. Added: "text" in label.
Changed: Position and label of "regular expressions" -> "advanced RegEx", because regular is ambiguous with ordinary to a layman. Added: tooltip to explain.