Add "returning user" test.
It should show the behaviour for when a user queues two albums, but with
a gap in between in which a couple of other people have queued albums.
The first users second album shouldn't pick up the priority of the first
one.

fixup: add basic tests for get_upload_candidate()