TestGetCandidate: Make sure time.time() always increases
time.time() is called every time that we mark that we have interacted with
someone (via the `user_update_times` dict) which is used in round robin
mode to choose the user who has been waiting for us for the longest
time.

In MSYS2 python seems to not defer to the system clock every time you call it
(at least in a VM, the python version says `Python 3.9.7 (default, Nov 21
2021, 22:02:56) [GCC 11.2.0 64 bit (AMD64)] on win32`). So if time.time() is
called quickly, like in unit tests, it can return the same value in subsequent
calls, despite the time being in microsecond resolution. This makes the round
robin code essentially fall back to picking the first user, the same as fifo
mode.

This situation (two or more users last interacted with at the "same" time) may
possibly occur in a real environment, but it isn't likely:

1. we interact with two users fast enough that the difference in time is
smaller than the resolution of the clock
    * unlikely to happen, a bit of this weirdness in the round robin code in a
      theoretical but rare case isn't the end of the world
2. the windows builds are running on a python with the same issue
    * the values returned by time.time() do differ if you don't call it really
      fast, so probably in a real situation we wouldn't be processing uploads
      so fast that it would be an issue. Probably

Regarding the patching, since this is the first occurrence in this codebase.
`@patch()` will return a mock of the named object to the methods under test,
with the expectation that the tests will set the mock up. With unittest, when
patch is used as a decorator, that means that the mock will be passed to every
test, which is a pain (more line noise in the tests, bigger diffs). Using the
`new` kwarg to patch means "here is my already setup mock, don't bother
passing a fresh one to the tests to set up".