searchfilters.ui: Add ! operator to size/bitrate
Removed: Phrase "and re-apply it", because the filters update by themselves when cleared. Added: Line break because paragraph was too long
Added: Not ("!") operator for File Size with example, for an opportunity to illustrate floating point values are supported
Added: Examples for < and > because this gives opportunity to illustrate the gibibyte unit
Removed: Irrelevant sentence about VBR files; changed to instead Added: Clarification about the only unit of measurement used is kb/s
Added: Not ("!") operator for Bitrate, add phrase "an operator", and change the word "and" to "or" to clarify only one operator is possible
Changed: Footer text to indicate the default filters are in the preferences, which is where this help text is displayed.