avoid using privileged ports
The port-range 1-1024 is reserved for privileged processes and can only
be opened by the administrative user. Of course there is nothing stopping
users from running Nicotine+ as administrative user, but since it's not too
hard to figure out which port a user is using, it's also not to hard to
find users who do run n+ as admin.