transfers.py: prevent incomplete filename from exceeding 255 bytes
e86229103aef8ee95027e391b5989eee5ea3441e fixed an issue where
downloading a file with a filename of 255 characters would fail
because a prefix was added to its filename in the incomplete directory
which would create a filename over 255 characters, which isn't possible
on many systems.
The restriction on most systems is 255 bytes, not characters, so this
could still fail when multi-byte characters are present. Convert
the filename to bytes when truncating (prefix doesn't need this because
its byte length is known).