transfers.py: revert "Update locally saved upload speed more often"
Since commit 66fce3fc02ae4095465555384528383a02162484, the upload speed sent in search results is updated every time an upload finishes. This is a bad idea, since the speeds can fluctuate a lot depending on the user we're uploading to. This also causes a discrepancy between the speed reported in search results compared to user info.