Clarify use of autoaway option, make it a button
The autoaway option could be disabled by setting it to 0, but this is
not extremely obvious. Split it into two options: UseAutoAway and
AutoAwayInterval. Create an away timer only if UseAutoAway is enabled.
Turn the label into a check button, clarify its functionality to the
user, and set the lowest possible value to 1 because the button now
disables it instead of when a value of 0 is used.