transfers.py: rework checks for offline user
Checking a new user's status immediately after starting an upload is not possible, since
the server hasn't sent the status yet. This resulted in an issue where uploads were
sometimes incorrectly marked as "User logged off" when attempting to start an upload,
and "Cancelled" as soon as we received the status from the server.

Rework the checks for offline users to use the result from the connection attempt
instead.