Restore compatibility with the legacy file transfer system
FileUploadInit and FileDownloadInit cannot be used to reliably determine if
a transfer is an upload or download, since the user who requests a download
initiates the file upload connection in the legacy file transfer system.

Also ensure that file transfers stuck at 0% "Transferring" can still time out
if they are stuck.