transfers.py: fix invalid handling of 'Queued' cancel reason
Soulseek NS uses the 'Queued' reason for rejecting an upload attept if
the download slot limit is reached. This resulted in Nicotine+
attempting to upload the file over and over until the transfer timeout
was reached.

We were also using the 'Queued' reason when someone attempted to upload
a file we didn't queue before, which is incorrect.