Purge all the frames (#2267)
After two years of work and 20 years of frames, it's time to purge all of them.

Fun fact: 'frame' refers to the wxFrame window widget in wxWidgets, which was used in PySoulSeek.

Another fun fact: In early 2020, the frame.py file contained 3500 lines. Two years later, after
the rename to mainwindow.py, it contains 1400 lines.