FolderContentsRequest: only use latin-1 when necessary
If someone shares a folder with e.g. Ã¶ in the path, encoding the path in latin-1
will give you ö instead. While this is the intended character, the path no longer
matches what the user shared, resulting in the folder not being found.

Only use latin-1 when no files are present in the FolderContentsResponse message.