slskproto.py: don't attempt to write data immediately
When establishing an outgoing connection, we attempted to write data on the
same iteration. Since outgoing messages are processed on the next iteration,
our outgoing message buffer was empty.

Additionally, remove special case for empty buffers, and always fail early
with an OSError exception when a write is unsuccessful. Fixes #2320