search.py: add context menu item to view user profiles
User profiles are an integral part of the Soulseek network. It's where you learn
more about a user's collection, requirements and how to get access to private files.

Currently, the menu item for viewing a user's profile is buried in a submenu,
making it awkward to access frequently. Add a primary menu item to make
user profiles more obvious and easier to access, similar to SoulseekQt.