Stop using 'Remote file error' transfer status
Soulseek NS only sends this status message in response to legacy download requests,
which we no longer use.

Modern Soulseek clients send a UploadFailed message when a file connection closed
unexpectedly, but Soulseek NS also sends it at the beginning of an upload attempt
when it can't read a file. Instead of marking these files with a "Remote file error"
status, which will be incorrect in most cases, add a new local "Connection closed" status.

Additionally, start sending a "File read error." rejection message when we can't read
a file when uploading, instead of a "File not shared." message. This is what
SoulseekQt does (although SoulseekQt sends an incorrect path, making the message
not appear in the UI).