CLI: prompt for username and password when necessary
These changes are enough to get you up and running in headless mode, but avoids
adding commands for logging in as another user or changing your password.

We should add these eventually, but it will require more discussion and work to
get them right. For now, I'm simply adding the bare minimum to get Nicotine+
3.3.0 out the door.