GUI: remove transfer tab highlights
In their current state, these are mostly visual noise rather
than being helpful. If you have a lot of transfers, the tabs
are almost always highlighted. Auto-cleared transfers are
not taken into account either.

If we bring back highlights in the future, we should be a lot
more restrictive about when to activate them.