fileproperties.py: remove download button
In its current form, it's only useful for single file downloads to the root
of the download folder, while in many cases you would expect to download the
whole folder. Remove the button for now to avoid confusion.

Perhaps we can add buttons to either download the file or whole folder in the
future, if people want it.