setup.cfg: stop using recursive globs
Using them caused data in folders without a __init__.py file to be
installed, which is not the behavior we want.

This behavior broke tray icons in the Flatpak package, since it always
attempted to use the tray icon files in the pynicotine package, which
are unavailable to the host system, instead of the icons exported to
the system.