Use lock icon for private search results
There were concerns that the lock icon could be confused with a
shopping bag in some icon themes, but I don't think the shield
icon we use as a replacement works well here. Switch to the lock
icon instead to see how it works out. Perhaps the icon tooltip
and greyed out rows makes it clear enough the results are private.